requirements.txt
uvicorn
sqlalchemy
apscheduler
fastapi-mail
loguru
pydantic
fastapi[standard]
pandas
numpy
python-dotenv
google-genai
google-generativeai
newspaper3k 
lxml_html_clean
psycopg2
